🚚 Move filters to separate files
