+++ Src/Notion.Client/Models/Filters/Checkbox.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class CheckboxFilter : SinglePropertyFilter
+    {
+        [JsonProperty("equals")]
+        public bool Equal { get; set; }
+
+        [JsonProperty("does_not_equal")]
+        public bool DoesNotEqual { get; set; }
+    }
+}
