+++ Src/Notion.Client/Models/Filters/Files.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class FilesFilter : SinglePropertyFilter
+    {
+        [JsonProperty("is_empty")]
+        public bool IsEmpty => true;
+
+        [JsonProperty("is_not_empty")]
+        public bool IsNotEmpty => true;
+    }
+}
